backend/requirements.txt
# AWS CDK and Infrastructure
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# FastAPI and Lambda
fastapi>=0.104.0
mangum>=0.17.0
aws-lambda-powertools>=2.30.0

# Database and Storage
boto3>=1.34.0
pydantic>=2.5.0

# Security and Validation
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0

# Load Testing
locust>=2.17.0

# Monitoring
watchtower>=3.0.0 